+++ ving/docs/change-log.md
@@ -24,6 +24,7 @@ outline: deep
 * NOTE: Eliminated SelectInput. Use Dropdown or FormInput type="select" instead.
 * Added prepend/append slots to FormInput for inserting icons into input groups.
 * Exposed new props for modifying the classes associated with PanelZone.
+* Added a coerce attribute to the FormInput component.
 
 ## 2024-04-29
 * Implemented: reformat page generator to use panel components on view and edit #139